+++ CHANGELOG.md
@@ -3,8 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
 ## [Unreleased]
-
 ### Fixed
+- `<FloatingActionList />` did not closed its listener event.
 - Prevent empty tooltips on `<NotAvailable/>` elements that contain a label but no description.
 
 ## [5.1.0] 2019-07-05